adding changes

+++ src/main/java/com/fabian/myphonebook/api/ContactController.java
@@ -1,14 +1,9 @@
 package com.fabian.myphonebook.api;
 
-
-
-import com.fabian.myphonebook.exception.ApiRequestException;
+import com.fabian.myphonebook.exception.ContactNotFoundException;
 import com.fabian.myphonebook.models.Contact;
 import com.fabian.myphonebook.services.ContactService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.lang.NonNull;
@@ -17,7 +12,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.util.List;
 
-/*@CrossOrigin("https://myphoneboook.herokuapp.com")*/
+
 @RequestMapping("api/phonebook/contact")
 @RestController
 public class ContactController {
@@ -46,14 +41,32 @@ public class ContactController {
     }
 
     @DeleteMapping(path = "{id}")
-    public void deleteContactById(@PathVariable("id") String id){
-        contactService.deleteContactById(id);
+    public ResponseEntity<Object> deleteContactById(@PathVariable("id") String id){
+
+        boolean contactExist = contactService.exists(id);
+        if(contactExist){
+            contactService.deleteContactById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else{
+            throw new ContactNotFoundException();
+        }
+
+
     }
 
     @PutMapping(path = "{id}")
-    public void updateContact(@PathVariable("id") String id,@Valid @NonNull @RequestBody Contact contactToUpdate ){
+    public ResponseEntity<Object> updateContact(@PathVariable("id") String id, @Valid @NonNull @RequestBody Contact contact ){
 
-           contactService.updateContactById(id,contactToUpdate);
+        boolean contactExist = contactService.exists(id);
+
+        if(contactExist){
+            contactService.updateContactById(id,contact);
+            return new ResponseEntity<>(contact, HttpStatus.OK);
+        }else{
+            throw new ContactNotFoundException();
+        }
+
+
 
 
     }
